+++ packages/streamotter/bin/streamotter.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { runProcess } from "@streamotter/cli";
+
+await runProcess();